Make the silence detect (aka VAD) less sensitive to noise at the low end.
When a user uses hardware muting the audio level is usually "digital silence" (-127dbOv) under normal circumstances. When an AEC is present there seems to be some residual signal being injected. This is still extremely low, but is still some 20dbOv which the algorithm detects as a significant jump, and therefore speech. However, 20dB from -127 to -107 is not nearly the same as -60 to -40, so when at that low end we add an extra 10% of "float" for the threshold.